+++ account/templates/account/change_list.html
@@ -1,0 +1,12 @@
+{% extends "admin/change_list.html" %}
+
+{% block object-tools-items %}
+    <li>
+        <form action="/admin/upload_csv/" method="post" enctype="multipart/form-data">
+            {% csrf_token %}
+            <input type="file" name="csv_file">
+            <button type="submit">Upload CSV</button>
+        </form>
+    </li>
+    {{ block.super }}
+{% endblock %}
