--- account/templates/admin/account/account/course_change_list.html
@@ -1,7 +1,0 @@
-{% extends "admin/change_list.html" %}
-{% block object-tools-items %}
-    <li>
-        <a href="{% url 'courses:upload_csv' %}" class="btn btn-primary">Upload CSV</a>
-    </li>
-    {{ block.super }}
-{% endblock %}
