--- account/templates/admin/upload_success.html
@@ -1,5 +1,0 @@
-{% extends "admin/base.html" %}
-{% block content %}
-  <h1>Upload CSV File</h1>
-  <p>CSV file has been uploaded successfully.</p>
-{% endblock %}
