+++ account/urls.py
@@ -1,16 +1,15 @@
 from django.contrib import admin
 from django.urls import path, include
-from . import views
+
 
 app_name = 'account'
 
-urlpatterns = [
-    # path('<int:course_id>/likes', views.like_course, name='like_course'),
-    # path('<int:course_id>/dislikes', views.dislike_course, name='dislike_course'),
-    path('<int:course_id>/likes/check',
-         views.check_course_like, name='check_course_like'),
-    # path('likes', views.user_wishlist, name='user_wishlist'),
-    path('admin/upload-csv/', views.upload_csv, name='upload_csv'),
-    # path('<int:course_id>/likes/count',
-    #      views.course_like_count, name='course_like_count')
-]
+urlpatterns = []
+
+# path('<int:course_id>/likes', views.like_course, name='like_course'),
+# path('<int:course_id>/dislikes', views.dislike_course, name='dislike_course'),
+# path('<int:course_id>/likes/check',
+#      views.check_course_like, name='check_course_like'),
+# path('likes', views.user_wishlist, name='user_wishlist'),
+# path('<int:course_id>/likes/count',
+#      views.course_like_count, name='course_like_count')
