+++ config/urls.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.urls import path, include
-import account
+import account.views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
