+++ courses/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class CoursesConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'courses'
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "courses"
