+++ courses/forms.py
@@ -1,0 +1,4 @@
+from django import forms
+
+class CSVUploadForm(forms.Form):
+    csv_file = forms.FileField()
