+++ courses/urls.py
@@ -8,6 +8,7 @@ urlpatterns = [
     path('', views.index, name='index'),
     path('all-courses', views.all_courses, name='all_courses'),
     path('<int:series_id>', views.series_detail, name='series_detail'),
+    path('admin/upload-csv/', views.upload_csv, name='upload_csv'),
     # path('<int:course_id>/likes', views.like_course, name='like_course'),
     # path('<int:course_id>/dislikes', views.dislike_course, name='dislike_course'),
     # path('<int:course_id>/likes/check', views.check_course_like, name='check_course_like'),
